remote: Relax the parsing logic even more
It turns out that git.git doesn't error when there isn't even an url entry in the config for the remote.

`git remote show` will indeed list a remote which has been configured as followed

    [remote "no-remote-url"]
        fetch =

This commit  changes the parsing logic of a remote, making every property optional. However, in order to be considered valid, a remote has to be declared in the config with, at least, one remote related property (even if it is empty).