remote: Relax the parsing logic MORE even more
`git remote show` will list a remote which has been configured as followed

    [remote "bogus-remote"]
        hocus = pocus

This commit makes the parsing logic of a remote even more lax. If no valid remote related config entry has been found, we will try to parse the config to find if there's at least one entry under the remote section. If that's not the case, the remote is considered not found.